+++ packages/zowe-explorer-api/__tests__/__unit__/vscode/ui/TableViewProvider.unit.test.ts
@@ -9,7 +9,7 @@
  *
  */
 
-import { EventEmitter, ExtensionContext, WebviewView } from "vscode";
+import { commands, EventEmitter, ExtensionContext, WebviewView } from "vscode";
 import { TableBuilder, TableViewProvider } from "../../../../src/vscode/ui";
 
 describe("TableViewProvider", () => {
@@ -45,8 +45,10 @@ describe("TableViewProvider", () => {
                     { apple: 9, banana: 10, orange: 11 },
                 ])
                 .build();
+            const executeCommandMock = jest.spyOn(commands, "executeCommand").mockImplementation();
             await TableViewProvider.getInstance().setTableView(tableOne);
             expect((TableViewProvider.getInstance() as any).tableView).toBe(tableOne);
+            expect(executeCommandMock).toHaveBeenCalledWith("setContext", "zowe.vscode-extension-for-zowe.showZoweResources", true);
 
             const disposeSpy = jest.spyOn(tableOne, "dispose");
 
@@ -55,6 +57,14 @@ describe("TableViewProvider", () => {
             await TableViewProvider.getInstance().setTableView(tableTwo);
             expect((TableViewProvider.getInstance() as any).tableView).toBe(tableTwo);
             expect(disposeSpy).toHaveBeenCalled();
+            executeCommandMock.mockRestore();
+        });
+        it("sets the table to null", async () => {
+            const executeCommandMock = jest.spyOn(commands, "executeCommand").mockImplementation();
+            await TableViewProvider.getInstance().setTableView(null);
+            expect((TableViewProvider.getInstance() as any).tableView).toBe(null);
+            expect(executeCommandMock).toHaveBeenCalledWith("setContext", "zowe.vscode-extension-for-zowe.showZoweResources", false);
+            executeCommandMock.mockRestore();
         });
     });
 
@@ -105,27 +115,6 @@ describe("TableViewProvider", () => {
                 { isCancellationRequested: false, onCancellationRequested: fakeEventEmitter.event }
             );
             expect(resolveForViewSpy).toHaveBeenCalled();
-        });
-        it("shows a placeholder if a table view is not assigned", async () => {
-            await TableViewProvider.getInstance().setTableView(null);
-            const fakeView = {
-                onDidDispose: jest.fn(),
-                show: jest.fn(),
-                viewType: "zowe.panel",
-                title: "SomeWebviewView",
-                webview: { asWebviewUri: jest.fn(), onDidReceiveMessage: jest.fn(), options: {}, html: "" },
-            } as unknown as WebviewView;
-            const htmlSetter = jest.fn();
-            Object.defineProperty(fakeView.webview, "html", {
-                set: htmlSetter,
-            });
-            const fakeEventEmitter = new EventEmitter<any>();
-            await TableViewProvider.getInstance().resolveWebviewView(
-                fakeView,
-                { state: undefined },
-                { isCancellationRequested: false, onCancellationRequested: fakeEventEmitter.event }
-            );
-            expect(htmlSetter).toHaveBeenCalledWith(TableViewProvider.FALLBACK_HTML);
         });
     });
 });