+++ packages/zowe-explorer/CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ### Bug fixes
 
-- Added placeholder text to the "Zowe Resources" panel to provide clarity when a table is not visible. [#3113](https://github.com/zowe/zowe-explorer-vscode/issues/3113)
+- The "Zowe Resources" panel is now hidden by default until it is used to display a table or other data. [#3113](https://github.com/zowe/zowe-explorer-vscode/issues/3113)
 
 ## `3.0.0-next.202409132122`